add color palette cycling, and improve sparkles
sparkles should now be more accurate to the behaviour of the original label badge it tries to emulate. that is, the left and right side of the badge both get 1 led lighting up at a time. on the original badge this is due to a hardware limitation, but here it's just a stylistic choice.